Refactor timetable utils into smaller files - The timetable utils file was originally 1400+ locs while its tests is another 1400. Refactoring them into smaller logical files makes it significantly easier to read, writes and test - I tried re-exporting symbols from utils/timetables but realized that it causes circular references which causes ReferenceError: can't access lexical declaration X before initialization
